requirements.txt
streamlit
pandas
numpy
plotly
requests
avanza-api
